tools/cli/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "scs-tools"
version = "0.1.0"
description = "CLI tools for Structured Context Specification (SCS) project scaffolding"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "SCS Contributors"}
]
keywords = ["scs", "structured-context", "project-scaffolding", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Code Generators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.0.0",
    "pyyaml>=6.0",
    "jinja2>=3.0.0",
]

[project.optional-dependencies]
validator = ["scs-validator>=0.1.0"]

[project.urls]
Homepage = "https://github.com/tim-mccrimmon/structured-context-spec"
Repository = "https://github.com/tim-mccrimmon/structured-context-spec"
Issues = "https://github.com/tim-mccrimmon/structured-context-spec/issues"

[project.scripts]
scs = "scs_tools.cli:cli"

[tool.setuptools]
packages = ["scs_tools", "scs_tools.commands", "scs_tools.templates", "scs_tools.utils"]

[tool.setuptools.package-data]
scs_tools = ["templates/**/*"]
